requirements.txt
django
Pillow
python-dotenv
django-unfold
gunicorn
